backend/pyproject.toml
[tool.poetry]
name = "dao-portal-backend"
version = "0.1.0"
description = "Backend API for DAO Portal"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = "^0.24.0"
sqlmodel = "0.0.8"
asyncpg = "^0.29.0"
pydantic = "1.10.13"  # Pinned exactly to v1.10.13
python-dotenv = "^1.0.0"  # Use this instead of pydantic-settings
celery = "^5.3.4"
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.1"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
gunicorn = "^21.2.0"
httpx = "^0.25.2"
prometheus-fastapi-instrumentator = "^6.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
ruff = "^0.1.6"
pytest-cov = "^4.1.0"
httpx = "^0.25.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"